backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cohere>=5.20.0",
    "fastapi>=0.123.9",
    "fastembed>=0.7.3",
    "markdown>=3.10",
    "openai-agents>=0.6.1",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.1.0",
    "qdrant-client>=1.16.1",
    "sentence-transformers>=5.1.2",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.1",
    "trafilatura>=2.0.0",
]
